refactor(agents): split bounded output capture
